cleanup;  patches and surface aggregation; tilemap disappearance fix.
Added patches to the draw surface lists. A rendering artifact has
emerged as a result, though, which is odd...

Also fixed the issue with the floor decals by removing the polygon
offsets. Most of the global surface parameters in the shaders are meant
for the actual BSP compiler, and not the renderer - I initially made the
mistake of thinking otherwise.