refactor/testing
did some incremental testing to aid in moving all of the textures over
to array memory.

Also refactored the bspVertex_t data type to a template struct to keep
the relationship between bspVertex_t and other derivatives of that type
in an isolated location. Currently there is only one template parameter:
the type which is used for tex coordinates (either glm::vec2 or
glm::vec3, in this case). the vec3 as a texture coord is necessary for
texture array sampling.